Move RequestPolicies to policy package. TokenCredentialsHandler -> Policy.
