+++ azure-client-runtime/src/main/java/com/microsoft/azure/AzureServiceClient.java
@@ -9,7 +9,7 @@ package com.microsoft.azure;
 import com.google.common.hash.Hashing;
 import com.microsoft.azure.credentials.AzureTokenCredentials;
 import com.microsoft.azure.serializer.AzureJacksonAdapter;
-import com.microsoft.azure.v2.AzureTokenCredentialsHandler;
+import com.microsoft.azure.v2.policy.AzureTokenCredentialsPolicy;
 import com.microsoft.rest.RestClient;
 import com.microsoft.rest.ServiceClient;
 import com.microsoft.rest.credentials.ServiceClientCredentials;
@@ -17,10 +17,10 @@ import com.microsoft.rest.v2.http.ChannelHandlerConfig;
 import com.microsoft.rest.v2.http.HttpClient;
 import com.microsoft.rest.v2.http.HttpRequest;
 import com.microsoft.rest.v2.http.HttpResponse;
-import com.microsoft.rest.v2.http.RequestPolicy;
-import com.microsoft.rest.v2.http.RequestPolicyChain;
+import com.microsoft.rest.v2.policy.RequestPolicy;
+import com.microsoft.rest.v2.policy.RequestPolicyChain;
 import com.microsoft.rest.v2.http.RxNettyClient;
-import com.microsoft.rest.v2.http.UseOtherHostPolicy;
+import com.microsoft.rest.v2.policy.UseOtherHostPolicy;
 import io.netty.channel.ChannelHandler;
 import okhttp3.OkHttpClient;
 import retrofit2.Retrofit;
@@ -63,16 +63,10 @@ public abstract class AzureServiceClient extends ServiceClient {
     protected AzureServiceClient(RestClient restClient) {
         super(restClient);
 
+        final AzureTokenCredentials credentials = (AzureTokenCredentials) restClient().credentials();
         // TODO: tests, refactoring-- less hacky way of getting credentials
-        final RxNettyClient rxnClient =
-                new RxNettyClient(new ChannelHandlerConfig(new Func0<ChannelHandler>() {
-                    @Override
-                    public ChannelHandler call() {
-                        return new AzureTokenCredentialsHandler((AzureTokenCredentials) AzureServiceClient.this.restClient().credentials());
-                    }
-                }, true));
-
-        rpHttpClient = new RequestPolicyChain(new UseOtherHostPolicy.Factory(), new RequestPolicy.Factory() {
+        final RxNettyClient rxnClient = new RxNettyClient();
+        rpHttpClient = new RequestPolicyChain(new AzureTokenCredentialsPolicy.Factory(credentials), new RequestPolicy.Factory() {
             @Override
             public RequestPolicy create(RequestPolicy next) {
                 return new RequestPolicy() {
