+++ client-runtime/src/main/java/com/microsoft/rest/v2/policy/RequestPolicy.java
@@ -4,8 +4,10 @@
  * license information.
  */
 
-package com.microsoft.rest.v2.http;
+package com.microsoft.rest.v2.policy;
 
+import com.microsoft.rest.v2.http.HttpRequest;
+import com.microsoft.rest.v2.http.HttpResponse;
 import rx.Single;
 
 /**
